Factor out postTimeEntriesHandler
